Make refreshtoken a separate app.
